Create genre_to_date_plotting_padina
This includes the json file with the instances of movies in each decade in each genre and the python file of the code that helped get that data. 
Soon to be added: python code for genre_to_date plots

Note: aside from the possible genres, the dictionary also includes a key called 'All' which includes the decade data for all movies rather than just the movies in specific genres.